Enhance Model Training: Eval Support + Mixed Precision (Part of #3)

+++ detection3d/core/lmk_det_infer.py
@@ -327,7 +327,7 @@ def detection(input_path, model_folder, gpu_id, return_landmark_file, save_landm
 
           # only keep probability of the largest connected component
           landmark_mask_cc = sitk.GetArrayFromImage(landmark_mask_cc)
-          masked_landmark_mask_prob = np.multiply(landmark_mask_cc.astype(np.float), landmark_mask_prob)
+          masked_landmark_mask_prob = np.multiply(landmark_mask_cc.astype(np.float32), landmark_mask_prob)
 
           # compute the weighted mass center of the probability map
           masked_landmark_mask_prob = sitk.GetImageFromArray(masked_landmark_mask_prob)
